option - progress bar only at main parts
