+++ .vscode/launch.json
@@ -10,7 +10,11 @@
             "request": "launch",
             "program": "${file}",
             "console": "integratedTerminal",
-            "justMyCode": true,
+            "justMyCode": true
+            // "internalConsoleOptions": "openOnSessionStart"
+            // "env": {
+            //     "QT_QPA_PLATFORM_PLUGIN_PATH": "C:/Users/Nir/Documents/GitHub/KagomePeriodicBP/.conda/Library/plugins"
+            // }
             // "pythonArgs": ["-Xfrozen_modules=off"]  //, "PYDEVD_DISABLE_FILE_VALIDATION=1"
         },
         {
