+++ configuration.json
@@ -1,6 +1,6 @@
 {
     "allow_visuals"   : true,
-    "debug_mode"      : true,
+    "debug_mode"      : false,
     "keep_logs"       : true,
     "save_files_With" : {
         "_comment" : "can be 'dill'/'pickle'/null ",
