+++ src/containers/__init__.py
@@ -1,6 +1,7 @@
 from .contractions import BubbleConConfig, MPSOrientation
 from .belief_propagation import BPConfig, BPStats, Message, MessageDictType
-from .imaginary_time_evolution import ITEConfig, ITEPerModeStats, ITESegmentStats, ITEProgressTracker, UpdateEdge, HamiltonianFuncAndInputs, UpdateEdgesOrder
+from .imaginary_time_evolution import ITEConfig, ITEPerModeStats, ITESegmentStats, UpdateEdge, HamiltonianFuncAndInputs, UpdateEdgesOrder
+from ._ite_tracker import ITEProgressTracker
 from .sizes_and_dimensions import TNDimensions
 from .global_config import Config
 from .results import MeasurementsOnUnitCell
