+++ src/physics/metrics/_negativity.py
@@ -2,7 +2,6 @@ import numpy as np
 import typing as typ
 from itertools import combinations
 from scipy.linalg import sqrtm
-import matplotlib.pyplot as plt
 # My modules:
 from physics.metrics.src.densitymats import DensityMatrix, CommonDensityMatrices
 from utils import assertions
